% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_download.R
\name{wu_download}
\alias{wu_download}
\title{Download a file with an option to unzip}
\usage{
wu_download(
  x,
  destdir = ".",
  auto_unzip = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A single filename or URL of a PNG or JPG image}

\item{destdir}{Destination directory}

\item{auto_unzip}{Automatically unzip the file, Logical}

\item{overwrite}{Overwrite existing files, Logical}

\item{quiet}{Suppress messages}
}
\value{
A character vector of the downloaded file. If downloading a zip file with auto_unzip = TRUE, it
return a vector of filename(s) where they were unzipped.
}
\description{
Download a file with an option to unzip
}
\details{
This will download a file (using \code{\link{download.file}} under the hood) with an
option to unzip it (if its a zip file that is).

\code{x} must be a valid URL using http or ftp. This function requires that the URL ends
in a filename (unfortunately not all URLs do).
}
