% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_gfr_elems.R
\name{wu_gfr_elems}
\alias{wu_gfr_elems}
\title{Generate elements for a Google Form Summary}
\usage{
wu_gfr_elems(resp_tbl, rpt_tbl, show_msg = TRUE)
}
\arguments{
\item{resp_tbl}{Responses tibble}

\item{rpt_tbl}{Report template}

\item{show_msg}{Show additional messages}
}
\description{
Generate elements for a Google Form Summary
}
