% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_gfr_report.R
\name{wu_gfr_report}
\alias{wu_gfr_report}
\title{Generate a complete Google Form Summary HTML Report}
\usage{
wu_gfr_report(
  gfr_elems,
  output_fn,
  output_dir = ".",
  rpt_title = "Google Form Summary",
  rpt_rmd = NULL
)
}
\arguments{
\item{gfr_elems}{Google Form Summary Elements object}

\item{output_fn}{Report filename}

\item{output_dir}{Output directory}

\item{rpt_title}{The title for the report}

\item{rpt_rmd}{The Rmd file for the report}
}
\description{
Generate a complete Google Form Summary HTML Report
}
\seealso{
\code{\link{wu_gfr_elems}}
}
