% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_gfr_snippet.R
\name{wu_gfr_snippet}
\alias{wu_gfr_snippet}
\title{Generate a snippet of HTML code for summary report}
\usage{
wu_gfr_snippet(gfr_elems, output_fn, output_dir = ".", overwrite = TRUE)
}
\arguments{
\item{gfr_elems}{Google Form Summary Elements object}

\item{output_fn}{Report filename}

\item{output_dir}{Output directory}

\item{overwrite}{Overwrite an existing file, Logical}
}
\description{
Generate a snippet of HTML code for summary report
}
\details{
This generates the same HTML content as \code{\link{wu_gfr_report}},
however only the report elements are saved to the HTML file. In other words
this function doesn't produce a full
}
\seealso{
\code{\link{wu_gfr_elems}}, \code{\link{wu_gfr_report}}
}
