% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_img_build.R
\name{wu_img_build}
\alias{wu_img_build}
\title{Generate a list of HTML tags to show stacked PNGs}
\usage{
wu_img_build(
  img_fns,
  display_first = TRUE,
  img_class = "stackme",
  center = FALSE,
  border = c("none", "1px solid gray")[1]
)
}
\arguments{
\item{img_fns}{A vector of image files names}

\item{display_first}{Display the first image initially}

\item{img_class}{The class that image tags will be assigned}

\item{center}{Center the DIV on the document}

\item{border}{A valid CSS style for the DIV border}
}
\description{
Generate a list of HTML tags to show stacked PNGs
}
\details{
This will return a list of HTML tags that when inserted into a R markdown HTML document
will display the PNG files as a build. You can use this function within code chunks
in R markdown to create a build effect from a set of PNG files. The PNG
files will be added to the DIV in the order received. All images should have the same
dimensions, and images after the first one should either have transparent backgrounds
or contain the earlier elements in them.

To use this function, add the following style to your Rmarkdown document:

\preformatted{
```{css echo = FALSE}
img.stackme {
  position:absolute;
  top:0;
  left:0;
}
```
}

Create the build with:

\preformatted{
```{r echo = FALSE, results = "asis"}
img_fns <- paste0("./images/rnotebooks (", 1:9, ").png")
htmltools::tagList(wrkshputils::wu_img_build(img_fns, display_first = TRUE, center = FALSE))
```
}
}
