% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_img_maketrans.R
\name{wu_img_maketrans}
\alias{wu_img_maketrans}
\title{Make PNG images transparent backgrounds}
\usage{
wu_img_maketrans(
  img_fn = NULL,
  img_dir = NULL,
  pattern = ".png$",
  ext_out = "_bgtrns.png"
)
}
\arguments{
\item{img_fn}{PNG filename}

\item{img_dir}{Image directory}

\item{pattern}{Pattern to match for input images}

\item{ext_out}{Output extension}
}
\description{
Make PNG images transparent backgrounds
}
\details{
This will take a directory of PNG files and create versions where the color white is transparent.
This can be useful if you want to use the images as a build.

You must pass
}
