% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_popup.R
\name{wu_popup}
\alias{wu_popup}
\title{Plot a png or jpg image file}
\usage{
wu_popup(x)
}
\arguments{
\item{x}{A filename or URL to a PNG or JPG image}
}
\description{
Plot a png or jpg image file
}
\details{
This will plot a PNG or JPG image. \code{x} can be either a local file or a
URL. This gives you the ability to create 'popup' slides in a R script that you
work through during a lesson.
}
