% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_gfr_elems.R
\name{wu_gfr_elems}
\alias{wu_gfr_elems}
\title{Generate elements for a Google Form Summary}
\usage{
wu_gfr_elems(resp_tbl, rpt_tbl, show_msg = TRUE)
}
\arguments{
\item{resp_tbl}{Responses tibble}

\item{rpt_tbl}{Report template}

\item{show_msg}{Show additional messages}
}
\description{
Generate elements for a Google Form Summary
}
\details{
This function takes two tibbles: i) the responses from a Google Form Survey
(or any kind of survey, where the format of the responses mirrors that of Google Forms),
and ii) a tibble that specifies how the questions should be summarized and presented
in the output report.

This function does not download any data from Google Sheets. It is presumed you have
already done that. It also doesn't do anything terribly useful with the summaries.
It is presumed you'll pass the object returned by this function to \code{wu_gfr_report()}, or a similar
function that renders or analyzes the results.

For an example of \code{rpt_tbl}, please see \href{http://www.google.com}{here}.
}
