% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_gfr_report.R
\name{wu_gfr_report}
\alias{wu_gfr_report}
\title{Generate elements for a Google Form Summary}
\usage{
wu_gfr_report(
  gfr_elems,
  output_fn,
  output_dir = ".",
  rpt_title = "Google Form Summary",
  report_rmd = NULL
)
}
\arguments{
\item{resp_tbl}{Responses tibble}

\item{rpt_tbl}{Report template}

\item{show_msg}{Show additional messages}
}
\description{
Generate elements for a Google Form Summary
}
