% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_img_maketrans.R
\name{wu_img_maketrans}
\alias{wu_img_maketrans}
\title{Make PNG images transparent backgrounds}
\usage{
wu_img_maketrans(
  x,
  pattern = ".png$",
  out_dir = NULL,
  out_fn = "img_\%03d.png",
  overwrite = FALSE
)
}
\arguments{
\item{x}{Vector of image file names or a directory}

\item{pattern}{Pattern to match if x is a directory}

\item{out_dir}{Output directory}

\item{out_fn}{A filename template}

\item{overwrite}{Overwrite existing files}
}
\description{
Make PNG images transparent backgrounds
}
\details{
This will take one or more PNG files (e.g., saved from PowerPoint slides) and create copies where white becomes
transparent. This can be useful if you want to use the images as a build.

\strong{TIP}: if the source of your PNG files are a series of PowerPoint slides (that you convert to PNG with
File >> Save As...), set the size of your PowerPoint slides to produce the desired output dimensions.
When you save PowerPoint slides to raster formats, it saves them as 96 dpi (unless you've tweaked that).
So for example if you want to create PNG files that are 1000 x 562 (roughly 16:9 that will display fine on
a single slide), set the size of your PowerPoint presentation to 10.4" x 5.9".

\code{x} can be a directory that contains image files, or a vector of filenames. If \code{x} is a directory,
you can use \code{pattern} to pass a regular expression.

\code{out_dir} is the output directory (optional). \code{out_fn} is a file name template that will be passed
to \code{sprintf} for evaluation. Example: \code{out_fn = \"slide_\%02d.png\"} will result in output files named
slide_01.png, slide_02.png, slide_03.png, etc. If \code{out_fn} is NULL, the original file names
appended with '_trnsbg.png' will be used for the output files.
}
\seealso{
\code{\link{wu_img_build}}
}
